+++ src/routes/post.js
@@ -2,35 +2,36 @@
 
 module.exports = (ctx)=>
 {
+    const router = require('express').Router();
     const limit = ctx.config.server.limit;
     const baseurl = ctx.config.server.api + '/post';
     const userSearchFields = {'_id':1,'following':1,'displayname':1,'image':1};
 
     // get items posted to public timeline
-    ctx.app.get( baseurl + '/public',(req,res) =>
+    router.get( baseurl + '/public',(req,res) =>
     {
         ctx.db.collection('posts').find({'repostid':{$exists: false, $eq: null}}) // only original post
                                   .limit(limit)
                                   .sort({'timestamp':-1}) // order by timestamp desc
                                   .toArray((err,data)=> {
-                if(err) return ctx.util.action.errorResult(err.message,req,res);
+                if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
                 const panelcolors = ['primary','default','success','info','warning','danger'];
                 const textcolors = ['white','red','blue','blue','red','blue'];
 
                 data.forEach((element,index)=> {
-                  element.text = ctx.util.string.bodyProcess(element.text); // process the body
-                  element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
-                  element.panelcolor = panelcolors[index % 6];
-                  element.textcolor = textcolors[index % 6];
+                    element.text = ctx.helper.string.bodyProcess(element.text); // process the body
+                    element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
+                    element.panelcolor = panelcolors[index % 6];
+                    element.textcolor = textcolors[index % 6];
                 });
 
-                ctx.util.action.jsonResult(req,res,data);
+                ctx.helper.action.jsonResult(req,res,data);
             });
     });
 
     // get items posted to user timeline
-    ctx.app.get( baseurl + '/home/:username',(req,res) =>
+    router.get( baseurl + '/home/:username',(req,res) =>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},{'fields':userSearchFields},(err,data)=>
         {
@@ -38,10 +39,10 @@ module.exports = (ctx)=>
                 data.following.push(data._id); // add my id for showing also my posts
                 ctx.db.collection('posts').find({'ownerid':{ $in:data.following}},{'limit':limit,'sort':{'timestamp':-1}})
                     .toArray((err,items)=> {
-                        if(err) return ctx.util.action.errorResult(err.message,req,res);
+                        if(err) return ctx.helper.action.errorResult(err.message,req,res);
                         items.forEach((element)=>
                         {
-                          element.text = ctx.util.string.bodyProcess(element.text); // process the body
+                          element.text = ctx.helper.string.bodyProcess(element.text); // process the body
                           if(element.repostid!==undefined) { // check if repost
                             element.isrepost = true;
                           }
@@ -51,31 +52,31 @@ module.exports = (ctx)=>
                           element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
                           //ctx.logger.debug(element);
                         });
-                        ctx.util.action.jsonResult(req,res,items);
+                        ctx.helper.action.jsonResult(req,res,items);
                     });
             }
             else {
-                ctx.util.action.okResult(req, res);
+                ctx.helper.action.okResult(req, res);
             }
         });
     });
 
     // get the items posted from an user
-    ctx.app.get( baseurl + '/:username',(req,res) =>
+    router.get( baseurl + '/:username',(req,res) =>
     {
         ctx.db.collection('posts').find({'username': req.params.username},{'sort':{'timestamp':-1}})
             .toArray((err,items)=> {
-                if(err) return ctx.util.action.errorResult(err.message,req,res);
-                ctx.util.action.jsonResult(req, res, items);
+                if(err) return ctx.helper.action.errorResult(err.message,req,res);
+                ctx.helper.action.jsonResult(req, res, items);
             });
     });
 
     // post a new message
-    ctx.app.post( baseurl,(req,res) =>
+    router.post( baseurl,(req,res) =>
     {
         ctx.db.collection('users').findOne({'username':req.body.username}, {'fields':userSearchFields},(err,data)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (data) {
                 var ObjectID = require('mongodb').ObjectID;
                 var newPost = {
@@ -89,47 +90,49 @@ module.exports = (ctx)=>
                 };
                 ctx.db.collection('posts').save(newPost,(err)=>
                 {
-                    if(err) return ctx.util.action.errorResult(err.message,req,res);
-                    ctx.util.action.okResult(req,res);
+                    if(err) return ctx.helper.action.errorResult(err.message,req,res);
+                    ctx.helper.action.okResult(req,res);
                 });
             }
         });
     });
 
     // post a new message
-    ctx.app.post( baseurl + '/repost',(req,res) =>
+    router.post( baseurl + '/repost',(req,res) =>
     {
         // check the user that wants to repost
         ctx.db.collection('users').findOne({'username':req.body.username},{'fields':userSearchFields},(err,user)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (user) {
-              const ObjectID = require('mongodb').ObjectID;
+                const ObjectID = require('mongodb').ObjectID;
 
-              // check the post to repost
-              ctx.db.collection('posts').findOne({'_id':req.body.id},(err,post)=>
-              {
-                if(err) return ctx.util.action.errorResult(err.message,req,res);
-                var newPost = {
-                    "_id": new ObjectID().toString(),
-                    "username": post.username,
-                    "ownerid": user._id,
-                    "repostid": post._id,
-                    "repostdisplayname": user.displayname,
-                    "repostusername": req.body.username,
-                    "displayname": post.displayname,
-                    "timestamp": new Date().toISOString(),
-                    "image": post.image,
-                    "text": post.text
-                };
-
-                ctx.db.collection('posts').save(newPost,(err)=>
+                // check the post to repost
+                ctx.db.collection('posts').findOne({'_id':req.body.id},(err,post)=>
                 {
-                    if(err) return ctx.util.action.errorResult(err.message,req,res);
-                    ctx.util.action.okResult(req,res);
+                    if(err) return ctx.helper.action.errorResult(err.message,req,res);
+                    var newPost = {
+                        "_id": new ObjectID().toString(),
+                        "username": post.username,
+                        "ownerid": user._id,
+                        "repostid": post._id,
+                        "repostdisplayname": user.displayname,
+                        "repostusername": req.body.username,
+                        "displayname": post.displayname,
+                        "timestamp": new Date().toISOString(),
+                        "image": post.image,
+                        "text": post.text
+                    };
+
+                    ctx.db.collection('posts').save(newPost,(err)=>
+                    {
+                        if(err) return ctx.helper.action.errorResult(err.message,req,res);
+                        ctx.helper.action.okResult(req,res);
+                    });
                 });
-              });
             }
         });
     });
+
+    return router;
 };