+++ src/routes/user.js
@@ -2,18 +2,19 @@
 
 module.exports = (ctx)=>
 {
+    const router = require('express').Router();
     const bcrypt = require('bcrypt');   
     const baseurl = ctx.config.server.api + '/user';
 
     // get the user using login and password
-    ctx.app.get( baseurl + '/authenticate/:username/:password',(req,res)=>
+    router.get( baseurl + '/authenticate/:username/:password',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},(err,data)=> {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (data) {
                 bcrypt.compare(req.params.password, data.password, (err, resbcrypt)=> {
                     if (resbcrypt === true) {
-                        ctx.util.action.jsonResult(req,res,{
+                        ctx.helper.action.jsonResult(req,res,{
                             "username": data.username,
                             "displayname": data.displayname,
                             "password": data.password,
@@ -25,23 +26,23 @@ module.exports = (ctx)=>
                             "imagepath": ctx.config.server.api + '/image/' + data.image
                         });
                     } else { // check password failed
-                        ctx.util.action.forbiddenResult(req,res);
+                        ctx.helper.action.forbiddenResult(req,res);
                     }
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     });
 
     // get the user using a username
-    ctx.app.get( baseurl + '/access/:username', (req,res)=>
+    router.get( baseurl + '/access/:username', (req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},(err, data)=>{
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (data) {
-                ctx.util.action.jsonResult(req,res,{
+                ctx.helper.action.jsonResult(req,res,{
                     "username": data.username,
                     "displayname": data.displayname,
                     "password": data.password,
@@ -54,67 +55,67 @@ module.exports = (ctx)=>
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     });
 
     // get the following of a user
-    ctx.app.get( baseurl + '/following/:username',(req,res)=>
+    router.get( baseurl + '/following/:username',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},(err, data)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (data) {
                 ctx.db.collection('users').find({'followers':data._id}).toArray((err,items)=> {
-                    if(err) return ctx.util.action.errorResult(err.message,req,res);
+                    if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
                     items.forEach((element)=> {
-                      element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
-                      //ctx.logger.debug(element);
+                        element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
+                        //ctx.logger.debug(element);
                     });
 
-                    ctx.util.action.jsonResult(req, res, items);
+                    ctx.helper.action.jsonResult(req, res, items);
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     });
 
     // get the followers of a user
-    ctx.app.get( baseurl + '/followers/:username',(req,res)=>
+    router.get( baseurl + '/followers/:username',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},(err,data)=> {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (data) {
                 ctx.db.collection('users').find({'following':data._id}).toArray((err,items)=> {
-                    if(err) return ctx.util.action.errorResult(err.message,req,res);
+                    if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
                     items.forEach((element)=> {
                       element.imagepath = ctx.config.server.api + '/image/' + element.image; // added image resource api
                       //ctx.logger.debug(element);
                     });
 
-                    ctx.util.action.jsonResult(req, res, items);
+                    ctx.helper.action.jsonResult(req, res, items);
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     });
 
     // sign up a new user
-    ctx.app.post( baseurl,(req,res)=>
+    router.post( baseurl,(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.body.username},(err,data)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
             if(data) { // username already existing
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
             else {
                 bcrypt.genSalt(10, (err,salt)=>
@@ -133,8 +134,8 @@ module.exports = (ctx)=>
                         };
 
                         ctx.db.collection('users').save(user,(err)=> {
-                            if(err) return ctx.util.action.errorResult(err.message,req,res);
-                            ctx.util.action.okResult(req,res);
+                            if(err) return ctx.helper.action.errorResult(err.message,req,res);
+                            ctx.helper.action.okResult(req,res);
                         });
 
                     });
@@ -144,15 +145,15 @@ module.exports = (ctx)=>
     });
 
     // follow an user
-    ctx.app.post( baseurl + '/follow',(req,res)=>
+    router.post( baseurl + '/follow',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.body.username1},(err,user1)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
             if (user1) {
                 ctx.db.collection('users').findOne({'username':req.body.username2},(err,user2)=>
                 {
-                    if(err) return ctx.util.action.errorResult(err.message,req,res);
+                    if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
                     // check if already follow
                     if(user2) {
@@ -166,7 +167,7 @@ module.exports = (ctx)=>
 
                         if( found ) {
                             ctx.logger.debug('[%s] already follow [%s]',user1.username,user2.username);
-                            ctx.util.action.forbiddenResult(req,res);
+                            ctx.helper.action.forbiddenResult(req,res);
                             return;
                         }
 
@@ -176,29 +177,29 @@ module.exports = (ctx)=>
                               ctx.db.collection('users')
                                     .updateOne({_id:user2._id},{$push: {followers:user1._id }});
                             });
-                        ctx.util.action.okResult(req,res);
+                        ctx.helper.action.okResult(req,res);
                     }
                     else {
-                        ctx.util.action.forbiddenResult(req,res);
+                        ctx.helper.action.forbiddenResult(req,res);
                     }
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     } );
 
     // follow an user
-    ctx.app.post( baseurl + '/unfollow',(req,res)=>
+    router.post( baseurl + '/unfollow',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.body.username1},(err,user1)=>
         {
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
             if (user1) {
                 ctx.db.collection('users').findOne({'username':req.body.username2},(err,user2)=>{
-                if(err) return ctx.util.action.errorResult(err.message,req,res);
+                if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
                 // check if already follow
                 if(user2) {
@@ -213,7 +214,7 @@ module.exports = (ctx)=>
 
                   if( !found ) {
                     ctx.logger.debug('[%s] does not follow [%s]',user1.username,user2.username);
-                    ctx.util.action.forbiddenResult(req,res);
+                    ctx.helper.action.forbiddenResult(req,res);
                     return;
                   }
 
@@ -223,27 +224,27 @@ module.exports = (ctx)=>
                           ctx.db.collection('users')
                                 .updateOne({_id:user2._id},{$pull: {followers: user1._id }});
                         });
-                  ctx.util.action.okResult(req,res);
+                  ctx.helper.action.okResult(req,res);
                 }
                 else {
-                    ctx.util.action.forbiddenResult(req,res);
+                    ctx.helper.action.forbiddenResult(req,res);
                 }
               });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     } );
 
     // get the user info
-    ctx.app.get( baseurl + '/info/:username',(req,res)=>
+    router.get( baseurl + '/info/:username',(req,res)=>
     {
         ctx.db.collection('users').findOne({'username':req.params.username},(err,data)=>{
-            if(err) return ctx.util.action.errorResult(err.message,req,res);
+            if(err) return ctx.helper.action.errorResult(err.message,req,res);
 
             if(data) {
-                ctx.util.action.jsonResult(req,res,{
+                ctx.helper.action.jsonResult(req,res,{
                     "username": data.username,
                     "displayname": data.displayname,
                     "email": data.email,
@@ -255,8 +256,10 @@ module.exports = (ctx)=>
                 });
             }
             else {
-                ctx.util.action.forbiddenResult(req,res);
+                ctx.helper.action.forbiddenResult(req,res);
             }
         });
     });
+
+    return router;
 };